requirements.txt
tqdm==4.65.0
transformers==4.28.1
datasets==2.11.0
huggingface-hub==0.13.4
accelerate==0.18.0

torch~=2.0.1
unstructured~=0.9.3
peft~=0.3.0
pdf2image
pdfminer
numpy~=1.23.5
paddlenlp~=2.5.2
uvicorn~=0.21.1
fastapi~=0.95.0
pydantic~=1.10.7
scikit-learn~=1.2.2
evaluate~=0.4.0
requests~=2.28.1
Pillow~=9.4.0